refactor(mock-consensus): 🔐 remove `Keyring`
previously, we added a `Keyring` newtype. this was a low-effort, thin
newtype wrapper around a `BTreeMap<VerificationKey, SigningKey>`.

this commit removes it.

NB: this is based on #4001.
